Fix graph connection pool handling
As we connected to graph before the actual fork of worker processes from
gunicorn master process, the connection pool was copied to multiple workers
(note we perform ping in the connection setup to verify correct connection
establishment so pool is not empty). This caused wrong connection handling on
worker processes which started using shared connection pool and unpredictable
errors in the deployment.